+++ src/surfaces/H2DE_texture.cpp
@@ -1,5 +1,6 @@
 #include "H2DE/surfaces/H2DE_texture.h"
 #include "H2DE/engine/H2DE_lerp_manager.h"
+#include "H2DE/engine/H2DE_error.h"
 
 // INIT
 H2DE_Texture::H2DE_Texture(H2DE_Engine* e, H2DE_Object* o, const H2DE_SurfaceData& sd, const H2DE_TextureData& td) : H2DE_Surface(e, o, sd), textureData(td) {